[fix] 게임 신청을 수락한 유저가 상대방의 이름 모르는 버그 수정

+++ Android/FunBox/app/src/main/java/com/rpg/funbox/presentation/map/MapFragment.kt
@@ -332,7 +332,7 @@ class MapFragment : BaseFragment<FragmentMapBinding>(R.layout.fragment_map), OnM
             val intent = Intent(context, GameActivity::class.java)
             intent.putExtra("StartGame", false)
             intent.putExtra("RoomId", applyGameServerData.roomId)
-            intent.putExtra("OtherUserId", applyGameServerData.userId)
+            intent.putExtra("OtherUserId", applyGameServerData.userId.toInt())
             startActivity(intent)
         }
